Move TPM related helpers to src/helpers
